refactoring, fixed chronicles
