+++ .changeset/cold-apricots-turn.md
@@ -1,0 +1,6 @@
+---
+"@equinor/mad-components": minor
+---
+
+BREAKING: 'Environment' component renamed to 'EnvironmentBanner', and no longer accepts props. Use
+EnvironmentProvider to provide the environment
