--- .changeset/mean-beds-joke.md
@@ -1,5 +1,0 @@
----
-"@equinor/mad-components": major
----
-
-BREAKING: add environment provider, Environment component no longer accepts props
