+++ packages/components/src/components/Environment/index.ts
@@ -1,3 +1,2 @@
 export * from "./EnvironmentBanner";
 export * from "./EnvironmentProvider";
-export * from "./EnvironmentBase";
