+++ packages/components/src/components/Environment/types.ts
@@ -1,0 +1,1 @@
+export type EnvironmentName = "dev" | "test" | "qa" | "prod";
